util module separated from each aug segments
