+++ farmer/ncc/augmentation/__init__.py
@@ -3,3 +3,4 @@ from .segmentation_aug import segmentation_aug
 from .classification_aug import classification_aug
 from .augmentation import augmentations
 from .augment_and_mix import augment_and_mix
+from .aug_utils import get_aug
